# Use CopyTo when resetting histogram in stats iterator